+++ korus-ui/components/LedaProvider/globalDefaultTheme.ts
@@ -39,6 +39,7 @@ import { defaultWizardTheme } from '../Wizard/theme';
 import { defaultMenuTheme } from '../Menu/theme';
 import { defaultCountDownTheme } from '../CountDown/theme';
 import { defaultLinkTreeTheme } from '../LinkTree/theme';
+import { defaultDrawerTheme } from '../Drawer/theme';
 
 export const globalDefaultTheme = {
   autoComplete: defaultAutoCompleteTheme,
@@ -49,6 +50,7 @@ export const globalDefaultTheme = {
   countDown: defaultCountDownTheme,
   dateTimeInput: defaultDateTimeInputTheme,
   dateTimeInputRange: defaultDateTimeInputRangeTheme,
+  drawer: defaultDrawerTheme,
   dropDown: defaultDropDownTheme,
   dropDownLink: defaultDropDownLinkTheme,
   dropDownSelect: defaultDropDownSelectTheme,