+++ storybook/storybook/stories/layout/DropDown/Examples/index.js
@@ -2,3 +2,5 @@ import componentSrc from '!raw-loader!@korus/leda/dist/components/Collapse/';
 
 export { componentSrc };
 export { Api } from './Api';
+export { Customization } from './Customization';
+export { Interaction } from './InteractionMode';