+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - LinkTree: Added new component
+- Drawer: Added new component
 
 ### Fixed
 - DropDownSelect: Fixed long strings display
+
 
 ## [0.42.0] - 2021-03-15